Move binary into type coercion
